# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentsfleet Control Plane API",
    "description": "API for managing workspaces, fleets, triggers, and runs.",
    "contact": {
      "name": "agentsfleet",
      "url": "https://agentsfleet.net"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.agentsfleet.net",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/v1/workspaces/{workspace_id}/fleets": {
      "get": {
        "tags": [
          "Fleets"
        ],
        "summary": "List fleets in a workspace",
        "description": "Returns visible fleets with the newest fleet first. Omit `starting_after` for the first page. Use the returned `next_cursor` to read the next page. ",
        "operationId": "list_fleets",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "`UUIDv7` of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "starting_after",
            "in": "query",
            "description": "Opaque page token from a prior response's `next_cursor` field, formatted `{created_at_epoch_ms}:{fleet_id}`. Omit for the first page. The retired `cursor` spelling is refused.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size. Defaults to 20. Clamped to a maximum of 100.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "post": {
        "tags": [
          "Fleets"
        ],
        "summary": "Create a fleet",
        "description": "Creates a fleet from one library entry. Use either `platform_library_id` or `tenant_library_id`. The service creates a default API trigger when the library has no trigger. ",
        "operationId": "create_fleet",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "`UUIDv7` of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallFleetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstalledFleetResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "424": {
            "description": "A dependency of this request failed"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/fleets/{fleet_id}": {
      "get": {
        "tags": [
          "Fleets"
        ],
        "summary": "Get a fleet",
        "description": "Returns one fleet's editable source, trigger markdown, bundle pin, trigger list, status, and lifetime counters. The response carries an `ETag` header over `source_markdown` and `trigger_markdown`. Send that value as `If-Match` when saving source changes to avoid overwriting another operator's edit. ",
        "operationId": "get_fleet",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "`UUIDv7` of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fleet_id",
            "in": "path",
            "description": "`UUIDv7` of the fleet.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "delete": {
        "tags": [
          "Fleets"
        ],
        "summary": "Permanently delete a fleet",
        "description": "Permanently deletes the fleet, events, runs, approvals, grants, keys, and memory. This cannot be undone. Set the fleet status to `killed` first. Otherwise the request returns 409. Requires an operator role. ",
        "operationId": "delete_fleet",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "`UUIDv7` of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fleet_id",
            "in": "path",
            "description": "`UUIDv7` of the fleet.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "patch": {
        "tags": [
          "Fleets"
        ],
        "summary": "Update a fleet",
        "description": "Updates fleet files or status. Every request field is optional. `config_json` and `trigger_markdown` cannot be used together. A killed fleet cannot be changed. Source editors can send `If-Match` with the `ETag` from GET. A stale tag returns 412 with the current tag in the problem body; omitting the header preserves last-write- wins behavior. Status changes require an operator role. Config changes require workspace membership. ",
        "operationId": "patch_fleet",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "`UUIDv7` of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fleet_id",
            "in": "path",
            "description": "`UUIDv7` of the fleet.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Optional source-version tag from GET. Stale values return 412 with the current `etag`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchedFleetResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "412": {
            "description": "The precondition the caller supplied does not hold"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FleetDetailResponse": {
        "type": "object",
        "description": "`GET /v1/workspaces/{workspace_id}/fleets/{fleet_id}` — one fleet, whole.\n\nThe list row's fields plus the editable surface. Flattened rather than\nnested under a `summary` key, because that is the shape the source editor\nalready reads.",
        "required": [
          "id",
          "name",
          "status",
          "source_markdown",
          "events_processed",
          "budget_used_nanos",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "budget_used_nanos": {
            "type": "integer",
            "format": "int64",
            "description": "Lifetime spend, in nanos."
          },
          "bundle_content_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "The bundle a runner materialises support files from."
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "When it was installed."
          },
          "events_processed": {
            "type": "integer",
            "format": "int64",
            "description": "Lifetime event count."
          },
          "id": {
            "type": "string",
            "description": "The fleet's identifier."
          },
          "name": {
            "type": "string",
            "description": "Its name in this workspace."
          },
          "source_markdown": {
            "type": "string",
            "description": "The authored `SKILL.md`, verbatim."
          },
          "status": {
            "type": "string",
            "description": "Where it stands in its life."
          },
          "trigger_markdown": {
            "type": [
              "string",
              "null"
            ],
            "description": "The authored `TRIGGER.md`, or `null` where the bundle carried none."
          },
          "triggers": {
            "type": [
              "object",
              "null"
            ],
            "description": "What may wake it, from the stored configuration.\n\n[`Triggers`] is `Option<Box<RawValue>>`: a stored document spliced\nthrough unparsed. `value_type` names the serialized shape, which is the\nonly thing a schema can say about it."
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "When it last changed."
          }
        }
      },
...
      "FleetSummary": {
        "type": "object",
        "description": "One fleet as a list page shows it.",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at",
          "events_processed",
          "budget_used_nanos"
        ],
        "properties": {
          "budget_used_nanos": {
            "type": "integer",
            "format": "int64",
            "description": "Lifetime spend, in nanos."
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "When it was installed."
          },
          "events_processed": {
            "type": "integer",
            "format": "int64",
            "description": "Lifetime event count. Server truth, never client arithmetic."
          },
          "id": {
            "type": "string",
            "description": "The fleet's identifier."
          },
          "name": {
            "type": "string",
            "description": "Its name in this workspace."
          },
          "status": {
            "type": "string",
            "description": "Where it stands in its life."
          },
          "triggers": {
            "type": [
              "object",
              "null"
            ],
            "description": "What may wake it, from the stored configuration.\n\n[`Triggers`] is `Option<Box<RawValue>>`: a stored document spliced\nthrough unparsed. `value_type` names the serialized shape, which is the\nonly thing a schema can say about it."
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "When it last changed."
          }
        }
      },
      "FleetsResponse": {
        "type": "object",
        "description": "`GET /v1/workspaces/{workspace_id}/fleets` — one page, newest first.",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FleetSummary"
            },
            "description": "The fleets on this page."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where the next page resumes, or `null` on the last one."
          },
          "total": {
            "type": "integer",
            "description": "How many are on this page — `list.zig` answers the page length, not the\nworkspace's whole count, and the name is the one that shipped.",
            "minimum": 0
          }
        }
      },
...
      "InstallFleetRequest": {
        "type": "object",
        "description": "`POST /v1/workspaces/{workspace_id}/fleets` — install one.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The operator's name for this instance, when they want one.\n\nThe same bundle backs many fleets in a workspace, each with its own name\nand its own webhooks. Absent means the bundle's declared name."
          },
          "platform_library_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "A published platform entry, by slug."
          },
          "tenant_library_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "This workspace's own entry, by identifier."
          }
        }
      },
      "InstalledFleetResponse": {
        "type": "object",
        "description": "What an install answers with.",
        "required": [
          "fleet_id",
          "name",
          "status",
          "webhook_urls"
        ],
        "properties": {
          "fleet_id": {
            "type": "string",
            "description": "The new fleet's identifier."
          },
          "name": {
            "type": "string",
            "description": "The name it was stored under — the drawn one, when a default collided."
          },
          "status": {
            "type": "string",
            "description": "Where it stands. `active`: the stream exists, so it is already leasable."
          },
          "webhook_urls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookUrl"
            },
            "description": "One entry per webhook trigger the bundle declared, keyed by source.\n\nAn empty object where the fleet declares none, never `null`: a client\niterating the map should not have to branch on its absence first."
          }
        }
      },
...
      "PatchedFleetResponse": {
        "oneOf": [
          {
            "type": "object",
            "description": "An empty body: nothing was written, and no row was even read.",
            "required": [
              "fleet_id"
            ],
            "properties": {
              "config_revision": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "Always `null` — the literal `@as(?i64, null)` the Zig answers, kept\nbecause a client distinguishes \"no write\" from a revision by it."
              },
              "fleet_id": {
                "type": "string",
                "description": "The fleet the request named."
              }
            }
          },
          {
            "type": "object",
            "description": "A configuration or source edit, with no transition asked for.",
            "required": [
              "fleet_id",
              "config_revision",
              "etag"
            ],
            "properties": {
              "config_revision": {
                "type": "integer",
                "format": "int64",
                "description": "The new revision — `updated_at`, echoed back by a conditional caller."
              },
              "etag": {
                "type": "string",
                "description": "The tag over the post-update source, for the editor's next save."
              },
              "fleet_id": {
                "type": "string",
                "description": "The fleet that changed."
              }
            }
          },
          {
            "type": "object",
            "description": "A transition, with or without an edit beside it.",
            "required": [
              "fleet_id",
              "status",
              "config_revision",
              "etag"
            ],
            "properties": {
              "config_revision": {
                "type": "integer",
                "format": "int64",
                "description": "The new revision."
              },
              "etag": {
                "type": "string",
                "description": "The tag over the post-update source."
              },
              "fleet_id": {
                "type": "string",
                "description": "The fleet that changed."
              },
              "status": {
                "type": "string",
                "description": "The status it now holds."
              }
            }
          }
        ],
        "description": "What a PATCH answers with — one of three shapes."
      },
...
      "WebhookUrl": {
        "type": "object",
        "description": "Where one declared webhook trigger is delivered.",
        "required": [
          "source",
          "url"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "The provider that sends it."
          },
          "url": {
            "type": "string",
            "description": "The address it posts to, on this deployment."
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Obtain a token via the CLI auth flow (POST /v1/auth/sessions) or GitHub OAuth"
      }
    }
  }
}